# docs/swagger.yaml
basePath: /
definitions:
...
  models.Address:
    properties:
      addressId:
        type: integer
      area:
        type: string
      city:
        type: string
      defaultadd:
        type: boolean
      district:
        type: string
      fullname:
        type: string
      housename:
        type: string
      landmark:
        type: string
      phone:
        type: integer
      pincode:
        type: integer
      state:
        type: string
      user:
        $ref: '#/definitions/models.User'
      userid:
        type: integer
    type: object
  models.User:
    properties:
      blockOrNot:
        type: boolean
      email:
        type: string
      first_name:
        maxLength: 50
        minLength: 2
        type: string
      id:
        type: integer
      is_admin:
        type: boolean
      last_name:
        maxLength: 50
        minLength: 1
        type: string
      otp:
        type: string
      password:
        type: string
      phone:
        type: string
      referal_code:
        type: string
      user_status:
        type: boolean
    required:
    - email
    - first_name
    - last_name
    - password
    - phone
    type: object
host: localhost:1111
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: An e-commerce site in Go using Gin framework
  title: E-commerce
  version: "1.0"
paths:
...
  /user/addaddress:
    post:
      consumes:
      - application/json
      description: Add a new address for the user
      parameters:
      - description: User address data
        in: body
        name: address
        required: true
        schema:
          $ref: '#/definitions/models.Address'
      produces:
      - application/json
      responses:
        "200":
          description: Address added successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerToken: []
      summary: Add User Address
      tags:
      - Add Address
      - Users
...
  /user/cancelorder:
    patch:
      consumes:
      - application/json
      description: Cancel an order by order ID
      parameters:
      - description: Order ID to cancel
        in: query
        name: order_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Order cancellation success message
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Order not found
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Cancel Order
      tags:
      - Orders
      - Users
      - ""
...
  /user/returnorder:
    put:
      consumes:
      - application/json
      description: Request to return an order by providing order ID and product ID
      parameters:
      - description: Product ID of the ordered product
        in: query
        name: product_id
        required: true
        type: integer
      - description: Order ID to be returned
        in: query
        name: order_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Order return success message
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Order not found
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Return Order
      tags:
      - Orders
      - Users
  /user/showorders:
    get:
      consumes:
      - application/json
      description: Get a list of user orders
      produces:
      - application/json
      responses:
        "200":
          description: User order details
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: No orders found
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Show User Orders
      tags:
      - Orders
      - Users
...
schemes:
- http
swagger: "2.0"